Store date and time separately and as text.
The way the form works has separate date and time fields, so we put that
in the model, and use methods to produce other formats and types for
the date time.

I cringe writing that, because I feel like the model shouldn't be driven
by the user interface.

This isn't working to/from the view, but the unit tests work.