Trying cookies for time zone.
The test case took a long time to figure out, because some documentation
led me to believe the 'option[checked]' was the selector for a drop-down.
But it turns out that you do indeed use 'option[selected]' for drop-downs.

This check-in doesn't include setting the cookie in the browser when
the user selects a time zone, nor does it disable the setting of the selected
item if there as a value in the cookie.