Further fixing the time zone handling.
Started to use cookies. Despite what some documentation seems to imply,
you need to use a string as the key in a cookie. Also fixed up the display
of the outages so they show as the user's time zone, regardless of how they
were entered. That's only for show so far, actually.