Defending against missing time zones
This is for issue #11, but not done and bigger.

I set some tests to make sure that the display/calendar views have
a time zone in the cookie. I fixed the tests to set a cookie when
I wanted the test to work.

I now need to do some Capybara tests to make sure my JavaScript is right
for the rest of the application.