Add time zone setting page.
Added a simple time zone controller that sets the time zone.
Did a test case for no cookie set. It sends the user to a
page to set the time zone, sets the cookie, then carries on.

The test case is sensitive to the time zone of the test
computer. Need to fix that.