Fixed wrong value in cookie in browser.
The JavaScript was setting the cookie in the browser to the browser's
time zone, overriding everything I was doing. Tests failed.
The webkit cookies are distinct from the Rails ones, and don't
seem to match, i.e. when integration testing, the cookies
aren't available as Rails -- or maybe I have to specify the session
or something.