Calendar movement working.
Wrote Capybara tests to move the calendar back and forth. In doing so,
learned about lambdas. The tests are dry.

There was a real trick with weeks: prev_week and next_week take you
to the start day of the week, which is not the same as weeks_ago and
weeks_since. I think the documentation implies that it should be the
same. It would be interesting to investigate whether I was just seeing
a phenomena of working Sunday morning -- but I doubt it. It was before
9:00 AM, but it's Sunday already UTC (time difference in the wrong
direction, in other words). Never mind.