Start day grid.
It looks like everyone just hard-codes the height of the cells
for the day grid. So far I just got a start on generating the
times. There are two days a year in many times zones where the
day view is strange, because of DST. I haven't handled that yet.